% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrt.R
\name{anova,jrt-method}
\alias{anova,jrt-method}
\title{anova method for objects returned by the \link{jrt} function.}
\usage{
\S4method{anova}{jrt}(object)
}
\arguments{
\item{object}{An object returned by \link{jrt}.}
}
\description{
anova method for objects returned by the \link{jrt} function.
}
