% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrt.R
\docType{class}
\name{jrt-class}
\alias{jrt-class}
\title{Object returned by the \link{jrt} function.}
\description{
Object returned by the \link{jrt} function.
}
\section{Slots}{

\describe{
\item{\code{input.data}}{The original data}

\item{\code{output.data}}{The output data with factor scores.}

\item{\code{fitted.model}}{The selected model.}

\item{\code{response.categories}}{The count of response categories.}

\item{\code{method.factor.scores}}{The method used to compute factor scores.}

\item{\code{imputed.data}}{The data with imputation.}

\item{\code{factor.scores}}{The factor scores with standard errors as a data.frame.}

\item{\code{factor.scores.vector}}{The factor scores as a vector.}

\item{\code{standard.errors.vector}}{The standard errors as a vector.}

\item{\code{mean.scores.vector}}{The mean scores as a vector.}

\item{\code{empirical.reliability}}{The empirical reliability.}

\item{\code{marginal.reliability}}{The marginal reliability.}

\item{\code{item.fit}}{Tests of item fit.}

\item{\code{person.fit}}{Tests of person fit.}

\item{\code{local.dependence}}{Tests of local dependence.}

\item{\code{sample.size}}{The sample size used in the model.}

\item{\code{number.of.judges.in.model}}{The number of judges (or items) in the model.}

\item{\code{column.names}}{The name used for the columns.}

\item{\code{mirt.object}}{The mirt object of the model.}
}}

