% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrt.R
\docType{data}
\name{ratings_missing}
\alias{ratings_missing}
\title{A simulated dataset with 350 cases judged by 5 judges, using a planned missingness design.}
\format{
A data.frame with 350 rows (cases) and 5 columns (judges):
\describe{
  \item{Judge_1}{Judgments of judge 1}
  \item{Judge_2}{Judgments of judge 2}
  \item{Judge_3}{Judgments of judge 3}
  \item{Judge_4}{Judgments of judge 4}
  \item{Judge_5}{Judgments of judge 5}
}
}
\usage{
ratings_missing
}
\description{
A simulated dataset with 350 cases judged by 5 judges, using a planned missingness design.
}
\keyword{datasets}
